Refactored tests to add pattern for mocking the proxy's upstream (#6)
no issue

Our test suite didn't yet have a pattern for intercepting requests sent to the upstream proxy, and making assertions on the results. Developing a pattern for that was the primary motivation for this change, in preparation for adding user agent parsing.

This commit changes our tests by adding a mock upstream server which is set as the PROXY_TARGET in the tests. The mock server is defined in test/testUtils/mock-upstream.js. It has a fallback route for all methods which returns 202 (the expected response from Tinybird for a successful request), and it logs all the requests to an array, which is available in each test so we can access the actual request sent to the upstream and make assertions on any part of the request.

A few other improvements to the test suite included in this commit:

- Changed from synchronous tests to asynchronous, so we don't have to explicitly call done() anymore
- Reorganized the test suite into more granular describe blocks
- Suppressed logging in the tests to make the output easier to read