Downgraded `ua-parser-js` library to 1.0.40 (#11)
ref
https://linear.app/ghost/issue/PROD-223/add-user-agent-parsing-to-the-proxy

- The UA parsing library we're using,
`[ua-parser-js](https://www.npmjs.com/package/ua-parser-js/v/1.0.40)`
switched to an AGPL license starting with version 2.0.0, which we cannot
use since this project is MIT licensed. Previous to version 2.0.0, this
library is licensed under an MIT license, which means we _can_ use it in
this project.
- This commit downgrades the `ua-parser-js` package from `2.0.3` to
`1.0.40` to remain in compliance with the license, and adds a renovate
rule to prevent automatically upgrading to `^2.x.x`