Refactored core validation and processing logic into a service (#12)
ref
https://linear.app/ghost/issue/PROD-223/add-user-agent-parsing-to-the-proxy

Currently this repo is only a simple HTTP proxy server, but eventually
it will need to operate in either "http" mode and/or "batch/queue" mode.
To make that transition easier in the future, we should decouple the
core business logic (validating and processing requests) from the
fastify HTTP proxy server.

This comes with the added benefit of making it easier to unit test each
step of the processing in isolation, so we can add all the edge cases in
fast unit tests, and reserve the E2E tests for high-level smoke tests to
cover the basic expected functionality.

This refactor is just a first step in this direction. There's still more
to do to fully decouple our core logic from the HTTP interface, but this
is a good start that's better than the current state, which has our core
logic closely coupled with the HTTP proxy endpoint.

This refactor introduces a new `proxy` service, composed of two macro
level steps: validation and processing. Validation is for checking that
the request is valid (e.g. checking that required query parameters are
present), while processing is for parsing and enriching the payload that
is ultimately sent to the upstream server.

This pattern makes it fairly simple to add other validation rules and
processing steps without much pain. We may add more macro steps in the
future (e.g. a final filtering stage after the processing is complete to
e.g. filter out bots by user agent).

Opportunities for further refactoring:
- The `proxy` service should ultimately be completely decoupled from the
`http` interface, but currently we are passing the `reply` object and
directly responding with any errors thrown. In the future, the `proxy`
service should just throw errors, and the HTTP server should then handle
those errors as appropriate, by e.g. responding 400 with an error
message.
- Factor the `filterQueryParams` utility into its own processor, and
remove the utils directory.
- Typescript would be nice, to define a strict interface for
`validators` and `processors`.
- The `proxy` service itself is difficult to test right now, because it
doesn't use dependency injection. We should probably refactor this into
a class with DI so we can test it more easily.
- I don't love that processors just change the request object in place —
in an ideal world they would be pure functions, but this seems to be the
standard pattern in Fastify — the request object itself gets mutated as
it passes through its lifecycle. Maybe an opportunity to improve this,
but it's not obvious to me how right now.