Improved UA parsing for browser, os and device (#16)
ref
https://linear.app/ghost/issue/PROD-223/add-user-agent-parsing-to-the-proxy

- Our UA parsing until now has just been logging the raw output from the
`ua-parser-js` library, which doesn't return the browser, os and device
in the exact format we want it.
- This commit adds some additional processing on top of what
`ua-parser-js` gives us. For example, it consolidates "Mobile Safari"
and "Safari" into just "Safari".
- It also adds some _very_ basic bot filtering based on a simple regex
test on the raw user agent string, similar to what we're doing in
Tinybird currently.
- For now, these values are still logged under `payload.meta`, so we can
gain more confidence with this running in a passive mode in production
with real world user agents.