Added url referrer processor (#38)
ref https://linear.app/ghost/issue/PROD-4
- added url referrer processing to match what is done in Ghost's member-attribution lib

We've added referrer parsing to the proxy such that we can do server-side parsing of the request payload. By doing this, we can match up the logic (via shared package in TryGhost/SDK) that parses urls for attributing source. We've put this in the meta field for now so we can run this in parallel with the existing referrer field - calculated on the frontend - which doesn't have as robust of matching, deduplication, etc.

The intent is to use the referrerSource field within Tinybird/traffic analytics. 

Co-authored-by: Chris Raible <chris@ghost.org>