Added support for running the analytics service in Ghost's docker compose network (#79)
no refs

This commit adds the ability to easily run the analytics service
side-by-side with Ghost running locally in Docker Compose. This provides
a more production-like setup when working on the analytics service.

- Ghost's docker compose setup now has a Caddy route handler to route
requests to `/.ghost/analytics` to this service.
- This commit adds a new `compose.ghost.yml`, which can be used in
addition to the base `compose.yml` to attach the analytics service to
the Docker network from the Ghost project.
- It also adds a convenience command: `yarn docker:dev:ghost` and some
documentation to the `README.md` for setting this up.

With this setup, Ghost's tracker script will route requests to
`http://localhost/.ghost/analytics/tb/web_analytics`. Ghost's Caddy
service will route those requests to the analytics service, which will
then forward the events to Tinybird. This is about as close to
dev/production parity as we'll be able to achieve.