Removed unused dependencies (#80)
no refs

We have some dependencies which aren't actually used in the repo. This
commit removes them from the `package.json` and `yarn.lock`.

There was also a line in the Dockerfile to install a different version
of Typescript — this shouldn't be necessary since typescript is already
declared as a dev dependency.