Added `salt-store` service using adapter pattern (#82)
refs
https://linear.app/ghost/issue/PROD-662/analytics-service-should-use-salt-hash-instead-of-storing-the-session

We want to start using a salt + hash fingerprinting method to uniquely
and anonymously identify "sessions", rather than persisting session IDs
client side in the tracker. This commit adds a `salt-store` service for
persisting the salt for each site in an arbitrary persistence layer. By
default it uses an in-memory store — this won't work in production since
it will be wiped on restarts, but it's sufficient for development and
prototyping.

Once we align on our persistence backend layer for Ghost(Pro) and for
self-hosters, we can create stores for them and swap them out easily
enough.