Added user signature service (#83)
refs
https://linear.app/ghost/issue/PROD-662/analytics-service-should-use-salt-hash-instead-of-storing-the-session

This commit adds a user signature service to the analytics service,
which generates an anonymous, unreversible user signature from a
request's IP address, user agent, and a unique salt per `site_uuid`. For
now it hasn't yet wired it up, this commit just adds the service with
unit tests.