Added user signature generation to proxy in shadow mode (#85)
refs
https://linear.app/ghost/issue/PROD-662/analytics-service-should-use-salt-hash-instead-of-storing-the-session

This commit wires up the user signature service to the proxy, and starts
generating the user signatures on each request. Currently it is logging
the generates signature to the `payload.meta.userSignature` field so we
can monitor this in production for a while before really turning it on.

Note: the salts for each `site_uuid` are stored in memory currently
since we don't have persistence setup yet. This means the salts can
rotate at any time, rather than the planned 24 hour rotation. We
shouldn't start using these signatures for real until we are able to
persist the salts.