Switched to default `ghost/ts` plugin from `eslint-plugin-ghost` (#87)
no refs

When we switched the project over to typescript, we added some custom
configuration that was needed because we weren't properly using the
`ghost/ts` plugin. This commit removes all the custom eslint config in
favor of the default `ghost/ts` plugin options.

The biggest change here is that it re-enables the `no-explicit-any`
rule, so we can't use `any` in our code anymore, which is IMO correct.
This change caught a few minor issues, like the `rewriteRequest` handler
not actually being used by the Fastify proxy, since it's not a valid
option.

It also adds an explicit `test:types` which runs typechecks, and is
included in the `yarn test` command, so it will always run in CI.