Added typechecking to test files (#89)
no refs

My IDE was flagging a lot of type errors in the tests with red
squiggles, but they weren't being checked in the `test:types` command,
as the tests were excluded. I hate looking at red squiggles, so this
commit enables type checking on the test files, but carves out an
exception for explicit `any`, since it is often required in tests (i.e.
when testing private methods).