Added databaseId parameter to FirestoreSaltStore (#92)
ref
https://linear.app/ghost/issue/PROD-662/analytics-service-should-use-salt-hash-instead-of-storing-the-session

The previous implementation would get the Firestore database based on
the GCP project ID, but both our staging and production DBs are in the
same project, so this won't work. This change allows you to specify a
database ID via environment variable, so we can point the analytics
service to the correct database.