Fixed the user signature generation (#93)
no refs

The user signature generation wasn't consistent for multiple requests
from the same browser because we were including a timestamp in the
signature, which makes it different for each request and defeats the
purpose of a user signature. This removes the timestamp, and also adds
the `site_uuid` itself to the signature to prevent cross-site tracking.