Reconfigured logging for structured logs in production (#99)
no refs

Our logs are currently always plain text, which makes the logs difficult
to reason about, especially in production. This commit changes the
logging config so it's still plain text in development to keep the logs
easily human readable, but changes to JSON logs in production to make
them more searchable.