Switched from vite to tsc and tsx (#102)
no refs

`vite` adds some unnecessary complexity to a project like this, which is
just a simple server-side node app. We're not using anything special
that requires `vite`, so this commit switches over to using the basic
`tsc` compiler and `tsx` for development. It also fixes hot reload in
Docker, so when you make changes to the code, the server will
automatically restart.