Added scheduler to cleanup old salts (#94)
ref
https://linear.app/ghost/issue/PROD-662/analytics-service-should-use-salt-hash-instead-of-storing-the-session

The user generator service recycles salts every 24 hours, but we
currently leave the old salts in the store. This commit adds a simpler
scheduler to delete any salts from before the current day UTC.

This can be disabled via config, and probably should be handled more
robustly at scale, but this works in development and should be fine for
self-hosters, since compose itself doesn't have a native scheduler.