Replaced session_id with salt + hash signature (#113)
ref
https://linear.app/ghost/issue/PROD-662/analytics-service-should-use-salt-hash-instead-of-storing-the-session

We've been logging the salt + hash based user signature for a while now
in "shadow mode", and everything seems to be working well. This commit
cuts it over into live mode, by replacing the `session_id` with the salt
+ hash based signature before sending the event to Tinybird. With this
live, we can change the tracker script in Ghost to stop sending the
`session_id` altogether.