Updated `yarn ship` command to work with protected main branch (#125)
no refs

Currently all commits merged to main are deployed to both staging and
production, which means we basically can't ever test anything on staging
before also deploying it to production. This has been fine during the
beta, but isn't a viable pattern as we move toward GA.

This commit is the first step toward separating the staging and
production deployments: all commits to main will still deploy to
staging, but they will only deploy to production if the version number
increases. This commit doesn't change the CI workflows yet, but it
replaces the current `yarn ship` command with a workflow that should
work with the protected main branch.

The `yarn ship` command does the following:

- Checks that you are on `main`, and that it's up to date
- Prompts for a version bump type (patch/minor/major)
- Checks out a new branch based on the new version
- Bumps the version in package.json and commits the change to the new
branch
- Outputs a link to open a PR from the branch
- Switches back to the main branch

A future commit will update the GHA workflows to tag the commit once the
PR with the version bump merges, and then deploy to production.