Fixed `ship.js` script rev-parse command error (#132)
no refs

The `yarn ship` script threw an error when trying to pull the latest
changes from `upstream/main` — this adds a fallback if `upstream/main`
(or `origin/main` if $GHOST_UPSTREAM isn't set) doesn't exist locally
for whatever reason. It also adds an eslint config for the scripts
directory and fixes the linter errors.