Updated check against upstream main (#134)
no refs

The step in the `yarn ship` script that checks if it's up to date with
main wasn't working consistently because it was checking the local
reference of upstream/main, which doesn't always update (at least not in
worktrees). This changes the check to directly check the head of the
remote, rather than the local reference to it.