Removed the environment matrix in favor of individual jobs (#142)
no refs

The previous workflow worked, but the "Deploy (production)" job in the
matrix would still run, even though it skips the actual deployment step.
The result is that Github Actions shows a big green checkmark next to
"Deploy (production)", even though it wasn't really deployed, which is
misleading and potentially scary.