Extracted logging hooks into a fastify plugin (#158)
We've been adding lots of functionality to the main `app.ts` file and
it's beginning to get a little hard to follow. In Fastify, it's a common
pattern to extract functionality into plugins, and register them in the
main app. This commit is the first move in this direction, pulling out
the logging hooks specifically into their own plugin.