Added `@tryghost/config` for configuration (#160)
This project has been using regular environment variables for all
configuration prior to now. To maintain consistency with our other
projects, we should aim to use established libraries when available.
This updates all uses of `process.env` to use `config.get()` instead,
except for specifically checking `NODE_ENV`.

For now it maintains backwards compatibility — all the config keys are
the same as the environment variables were, which means all caps and no
nesting. We might want to change this eventually to take advantage of
nested config, but that will require some changes to our
staging/production environments as well. For now this should be
sufficient.