Fixed logging trace correlation for GCP (#162)
It's currently difficult/impossible to correlate logs to a particular
request. This should pick up the GCP trace headers so we can drill down
from a request to the logs that were generated while processing it. GCP
keeps this trace throughout their whole serverless system (i.e. on
pub/sub events created), so this will be really useful to track events
through the system.