Reverted change to url logging (#166)
This isn't working as well as I'd hoped — it's using `main.ghost.org` as
the host, but it's omitting the `/.ghost/analytics` part of the path
since we strip that in the API Gateway. This means the URL that we put
together doesn't actually exist, and could lead to confusion.

This just reverts the change so we'll continue to log the `request.url`,
without the host and protocol.