compose port automation (#170)
When working with multiple worktrees, we need to avoid port conflicts.
This is done by setting the ports in a .env file manually for each
worktree. This is tedious to do manually, so this script helps automate
it.

Running `./scripts/setup-worktree.sh {OFFSET}` will generate a new .env
file, which calculates the appropriate ports for each service based on
the provided integer OFFSET, and set the COMPOSE_PROJECT_NAME
appropriately to avoid conflicts with other worktrees.