Added basic worker script that will eventually be the batch processor (#172)
ref
https://linear.app/ghost/issue/PROD-661/analytics-service-should-batch-events-when-sending-to-tinybird

We plan to have a batch processor subscribe to our pub/sub topic and
process the events in batches, then forward them to Tinybird. This sets
up an initial worker thread which currently does nothing other than log
a heartbeat, so we can create the infrastructure in GCP to run the
worker, then iterate from there.