Added basic synthetic monitoring / healthchecks suite (#169)
At the moment, we don't have any alerting or checks that confirm event
ingestion is working fully from end-to-end, from a Ghost site's frontend
through to the analytics service and to Tinybird.

This playwright based test suite is meant to run against live sites in
staging and production to check that event ingestion is running. The
idea is to run this on a schedule in Github Actions, and fire an alert
if event ingestion fails.