Added worker process (#177)
ref
https://linear.app/ghost/issue/PROD-661/analytics-service-should-batch-events-when-sending-to-tinybird

This commit adds a basic worker process which just logs a heartbeat on a
loop. This will be the entrypoint for our batch processor, which will
run in a separate Cloud Run instance in production, and as a separate
worker in docker compose for local development and/or self hosters.