Removed the duplicate build for the worker (#187)
ref
https://linear.app/ghost/issue/PROD-661/analytics-service-should-batch-events-when-sending-to-tinybird

I've learned that our worker process will need at least a simple
healthcheck endpoint in order to run properly in GCP Cloud Run, which
means it will need to use Fastify. With this learning, I've realized we
don't actually need two separate entrypoints for the server and the
worker — we can _always_ run the `server.ts` entrypoint, and
conditionally load the main `app.ts` or a new `worker-app.ts` depending
on an environment variable.

This commit resets our build pipeline to something closer to what we had
previously, with a single vite.config.ts and a single build command. It
removes the worker process entirely. After this we can add the
`worker-app.ts` file and the conditional logic in `server.ts` to run the
appropriate app based on an environment variable.