Added a simple "black box" e2e test (#205)
Our current testing approach with integrations tests is overly complex.
We don't currently have a simple test that simply makes a request to the
service, and makes assertions that it receives the correct response.
This is an initial test to do just that, so we can quickly verify that
the service is running. A follow up commit will set this test up to run
in CI, along with the other existing tests.