Added minimal Tinybird events API client (#215)
ref
https://linear.app/ghost/issue/PROD-661/analytics-service-should-batch-events-when-sending-to-tinybird

Now that we're not just proxying requests more-or-less as is to
Tinybird, we'll need a proper API client to send batches of events to
Tinybird. This adds the bare minimum needed to POST events to Tinybird's
events API endpoint — will need to do some additional work in here for
error handling and potentially passing additional parameters, but this
is a solid foundation to build on.