Added JSON Schema based validation to proxy endpoint (#216)
ref
https://linear.app/ghost/issue/PROD-661/analytics-service-should-batch-events-when-sending-to-tinybird

For the events we publish to pub/sub, we'll want to have cleanly defined
schemas so that consuming services can understand what to expect in the
message bodies. We've also seen a handful of "User agent not found"
errors in production, due to improper/missing validation.

This commit adds an explicit JSON schema for the incoming requests so we
can quickly validate the headers, query string and body of each incoming
request, and immediately return a 400 error if the request fails
validation. It also creates a framework for managing our event schemas
as different stages of the process.