Updated validation function to log errors before responding (#219)
Our request validation is working correctly so far, but it's not logging
the validation errors, which makes it difficult to debug if e.g.
healthchecks fail due to a validation error. This commit adds some
simple logging when request validation fails.

It also runs the validation on the body, headers and query string before
responding, rather than failing fast and returning early, so we can see
all validation errors in a request at once.