Refactored proxy endpoint to use `@fastify/reply-from` instead of `@fastify/http-proxy` (#223)
The `@fastify/http-proxy` plugin we were using was overly complex for
our needs — it's designed to be used as a fully fledged proxy server for
handling multiple routes with complex logic. It's ultimately built on
`@fastify/reply-from`, which is a plugin for forwarding a single
request/endpoint to an upstream server.

Really all we need is `@fastify/reply-from`, and this also gives us more
control over the request handling — this will make it a lot easier to
conditionally proxy the request or simply publish the event to pub/sub.
It also allows us to take advantage of Fastify's built in request
validation using JSON schema since we are now defining our route
directory.