Added wiremock for e2e testing (#226)
Our current e2e tests we're not fully testing that we are sending the
right data to Tinybird. This commit sets up wiremock in place of our
`/local-proxy` endpoint, so we can not only check the response we get
from our own fake endpoint, but also make assertions on the data we're
sending to the `PROXY_TARGET`.