Removed referrer data from meta field
ref https://linear.app/ghost/issue/PROD-2071/

I've pulled out the .meta.referrer{x} data and am returning the referrerSource as the payload.referrer.

While we use the url and medium for processing logic, those values are irrelevant to display on the Tinybird side, and we don't need to pass them along (unless we're curious for auditing).