Refactored compose files to improve parity between local and CI (#233)
Currently our main `compose.yml` and our `compose.ci.yml` are two
completely different files, which is a pain to maintain and can lead to
drift between our local dev environment and our CI environment. This
changes to using a base `compose.yml` for both environments, with a
`compose.override.yml` for local development (to add bind mounts and
volumes for local development) and a `compose.ci.yml` which only
includes overrides for the CI environment.

This should help prevent issues where tests pass locally but not in CI,
and make it easier to maintain these two environments, which are 99%
shared configuration.