Added e2e tests to CI (#232)
Currently our e2e tests only run locally against the services running in
docker compose. In CI, we don't ever install the node dependencies
except in the Docker image itself, so in order to run the e2e tests in
CI we'll need to modify the e2e tests to be able to run in docker
compose.

This adds a new e2e-test service to our `compose.yml` file and makes the
required changes to the e2e test suite to be able to run in this
environment. This way we can run the e2e tests locally with `docker
compose run --rm e2e-test` (or `yarn docker:test:e2e`). It also paves
the way to running these tests in CI.