Added `page-hits-raw schema` for pub/sub messages (#236)
ref
https://linear.app/ghost/issue/PROD-661/analytics-service-should-batch-events-when-sending-to-tinybird

We're currently publishing messages to pub/sub for all incoming requests
that pass validation. However, we're not passing all the data that we
need to in the message, and we're also passing some data that we
definitely don't need. This commit adds a JSON Schema for the
`page-hits-raw` events, which we can use to validate the messages before
publishing to the topic, and when receiving them from the topic.

Next up: change the payload that we're actually publishing to pub/sub,
according to this schema.