Updated publisher to use `page-hit-raw` schema when publishing (#239)
Currently the payload that is included in the `page-hits-raw` published
event is incorrect — it doesn't have all the fields we need, and it
includes a few fields we don't need. This updated the payload of the
event we're publishing to use the `page-hit-raw` schema, which includes
all the fields we need to process the event downstream.