Added `BatchWorker` class and message validation (#243)
ref
https://linear.app/ghost/issue/PROD-661/analytics-service-should-batch-events-when-sending-to-tinybird

To simplify testing, this pulls out the `BatchWorker` functionality from
the Fastify worker-plugin. The worker-plugin now just handles creating,
starting and stopping the BatchWorker, so we can test the BatchWorker
itself easily in isolation.

It also adds message validation with simple logging for now to indicate
if the message was valid or not. They should _all_ be valid since we're
validating them before publishing them, but it doesn't hurt to
doublecheck.