Added page-hit-processed schema and transforms (#245)
ref
https://linear.app/ghost/issue/PROD-661/analytics-service-should-batch-events-when-sending-to-tinybird

We're currently receiving and validating the raw page hit messages in
the worker process. This adds a page-hit-processed schema, including
transform functions to accept a pageHitRaw and output a
pageHitProcessed.

It adds the processing to the BatchWorker and logs the result for now.
Next step is to send these events to the `analytics_events_test`
datasource in Tinybird, then start batching them.