Added Tinybird POST request for each event in the worker (#255)
ref
https://linear.app/ghost/issue/PROD-661/analytics-service-should-batch-events-when-sending-to-tinybird

This commit begins forwarding the events from the worker to Tinybird,
specifically to the `analytics_events_test` datasource, so we can start
to validate that the batching pipeline delivers the same results as the
proxy pipeline.

This is the final incremental step before adding the actual batching
logic to the worker, which will accumulate events before sending them
all in a batch to Tinybird. This incremental step is mainly to confirm
that the Tinybird client for sending events to Tinybird is working as
expected.