Fixed referrer validation failing (#261)
The `ghost-stats` script no longer sends the `referrer` field directly,
instead sending a `parsedReferrer` object. This service treats the
referrer as a required value, and therefore any requests from the
`ghost-stats` script are failing the validation, and thus not ending up
in Tinybird.

This commit makes the referrer field optional, and adds an optional
`parsedReferrer` field, so that these requests should pass validation
and succeed.