Updated compose services to use Firestore for salt storage (#264)
The service and the worker were both defaulting to the memory store when
running locally, resulting in mismatching user signatures. This isn't
really a problem since we use Firestore in production, but it did give
me a scare. This points both services to use the Firestore emulator in
docker compose for salt storage, so the user signatures match from each
service for the same request.