Added user-agent to page-hit-processed schema (#265)
Currently the proxy path is sending along the `user-agent` field in the
payload, while the batching path is not. This adds the user-agent to the
page-hits-processed schema so that it's included in the payload that's
sent to Tinybird. We will likely want to remove this soon, but for now
this keeps the two paths in sync so we can validate that they are the
same.