Added e2e test to ensure the batch and proxy mode deliver the same event (#267)
For this period where we're running both the proxy mode and batch mode
in parallel, we need to make sure that the processed events are exactly
the same. This e2e test enforces this, by checking the the event hitting
`analytics_events` and `analytics_events_test` are strictly equal.

Eventually we won't need this test, as they requests will go through the
same processing pipeline, but for now there are too many differences
between the two pipelines to be confident in this without an e2e test to
verify it.