Fixed batch worker throwing error when POST request fails (#268)
When a POST request to Tinybird from the batch worker fails (for
whatever reason, ie `EHOSTUNREACH` errors), the batch worker is catching
the error, logging it then rethrowing it and crashing. These transient
errors will always happen, so we definitely don't want to rethrow this
error and crash.

This removes the rethrow — we're nacking the message so it will be
retried, and this allows the worker to keep processing messages without
interruption.