Added batching logic to the `BatchWorker` (#270)
ref
https://linear.app/ghost/issue/PROD-661/analytics-service-should-batch-events-when-sending-to-tinybird

This commit adds batching logic to the BatchWorker, which until now has
sent new events as it receives them. Now it will process events as it
receives them and add them to an in-memory queue. It will flush the
batch (post it to Tinybird) every 1 second, or when the batch size
reaches 50. This should allow the single batch worker to scale to a
theoretical 5k events per second before hitting Tinybird's rate limit,
which gives us _plenty_ of breathing room for now.

It will wait until the messages are successfully sent to Tinybird before
acking them, and if the post fails, it will nack all the messages so
they'll all be retried.