Renamed `incoming-event-request` schema to `page-hit-request` schema (#277)
Currently the endpoint we're serving the proxy on is
`/tb/web_analytics`. Soon we'll be renaming this to something like
`/api/v1/page_hit` to make it more explicit what it's for. In the future
we may have other event types with their own endpoints and schemas, so
`incoming-event-request` feels far too generic.