Added tracker token override if provided in environment (#281)
ref
https://linear.app/ghost/issue/PROD-1961/remove-the-tracker-token-from-ghost-config-and-tracker-script

We ultimately don't want the tracker token to come from Ghost, since
this exposes it in the browser. This commit overrides the tracker token
in the request by stripping the existing token from the url query
parameters if provided, and adding the `TINYBIRD_TRACKER_TOKEN` from
environment as an authorization bearer token.