Improved assertions in e2e tests and removed unnecessary tests (#284)
The assertions in the e2e tests were too generic, using `url.includes()`
to check the passed query parameters. This isn't specific enough to
catch the difference between `analytics_events` and
`analytics_events_test`, and was resulting in false failures. This
cleans that up, and also adds a default value to the test token in case
a token isn't provided via environment variable.