Added more thorough logging for validation errors (#285)
We're seeing an increase in validation errors over the last few days,
but our logs don't show the reason for the errors. We should relax the
validation rules to be much more liberal in what we accept, but in order
to do so we also need more visibility into what we're currently
rejecting.