Inlined the processRequest method into the route handler
This is a step towards simplifying the logic in the proxy plugin. Currently the pub/sub publishing and the request processing are too tightly coupled — we did it this way initially to test the proxying behavior in parallel to the batching behavior, but now we need to be able to choose one or the other.

The processRequest method is now inlined into the route handler, so we can start to untangle the logic in one place.