Improved validation test specificity (#289)
Our validation tests in `app.test.ts` were checking that we receive 400,
but not what the actual error was. This means the tests were passing,
even though we were receiving 400 for the wrong reasons. This updates
the validation tests to check the error message in addition to the
status code, so we can be more confident in our tests.

A good example: we actually _do_ allow requests without a token, but
since the "should reject requests without a token" test was just
checking the status code, and it was failing because the body was empty,
this test was (incorrectly) passing.