Added request body logging for all validation errors (#291)
We're been seeing an increased number of validation errors over the last
few days, but it's sometimes difficult to tell if we really should be
rejecting these requests or not, because the logs don't include the
request body. This adds the request body to the log's payload, so we can
evaluate if the request really should be rejected or not.