Split publishing and proxying events into their own handlers (#294)
Currently we are _both_ proxying requests inline and publishing the
events to Pub/Sub for our batch worker to process in parallel. We need
to be able to configure this, so it does one or the other, but not both.

This is a first step in that direction, by clearly splitting the
handler's logic for each strategy. After validating this, we can then
add some conditional logic based on configuration to use either the
inline strategy or the batching strategy, but not both.