Readded parsed referrer to payload (#300)
The `parsedReferrer` values are exactly what was sent to the analytics
service. Even though we parse the values out into `referrerUrl`,
`referrerSource`, `referrerMedium`, it could prove useful to have the
original values sent to the service in the payload in the future for
debugging. This adds the `parsedReferrer` values back to the payload if
they were provided.