Aligned the proxy processing pipeline with the batch pipeline (#303)
The direct proxy and the batch pipeline are currently using slightly
different paths when processing the requests. Now that I've sorted
through all the differences, and the events landing in each table are
perfect matches, we can start using the more robust processing pipeline
built for the batch worker using TypeBox and JSON Schemas for the proxy
processing as well. This will guarantee that the processed events will
always be the same, and reduces the temporary duplication we've had in
the code while running each of these in parallel.