Switched to using batch mode only if pub/sub topic is configured (#305)
Now that the pipelines are confidently publishing the same data to
Tinybird, we should stop running both in parallel and switch over
completely to the batch worker. This commit makes the strategy
configurable, currently based on whether the pub/sub topic environment
variable is configured. We'll probably want to change this to be more
explicit eventually, but for now it meets our needs just fine.