Fixed flaky e2e tests in CI (#312)
Since switching to batch mode by default in our e2e tests, they've
become a bit flakier due to the more variable timing of when the request
hits wiremock. These tests had a constant 1 second timeout to wait for
the request to be received by wiremock, but even that often wasn't
enough, and made the tests unnecessarily slow.

This fixes the flakiness by polling wiremock for the request every
100ms, up to a 10 second timeout. As a result, the tests are more
reliable, and often faster as well.