Added docker service and e2e test for analytics-service-proxy (#316)
Now that we're not _both_ proxying and batching requests, we need a way
to easily toggle between proxy and batch mode when developing locally,
and in our e2e tests. This adds a dedicated service for proxy mode, and
an e2e test for it.

It also removes the unnecessary validation e2e tests, since these are
covered with faster local unit/integration tests.