Relaxed validation rules on payload.href (#326)
The validation on the `payload.href` field is seemingly too strict, as
it's returning validation errors for values that, at least to my eye,
are valid URLs. Maybe they are _technicaly_ invalid in some way, but I
don't think we should be rejecting them. This removes the URL validation
on the href field in favor of simpler `String()` validation.