Added OpenTelemetry tracing and GCP Cloud Trace exporter (#331)
We're seeing errors in GCP that are difficult to explain/debug without
having better monitoring and observability tools. This adds
OpenTelemetry tracing so we can see what's going on in the app and debug
performance issues.

For local development, it adds a Jaeger service to our docker compose
setup. The service and worker will both export traces to Jaeger when
running locally. When in production, the `K_SERVICE` environment
variable will be set by Google Cloud Run, which will then tell the app
to export traces to Google Cloud Trace.

For now this just uses the default node auto instrumentations, but once
we get this working we can start to create our own spans.