Added File based Salt Store for self hosting configuration (#357)
ref
https://linear.app/ghost/issue/PROD-2300/add-file-based-salt-store-for-self-hosters

- The analytics service requires some kind of persistent storage for the
salts that are used as part of the user signature generation. The salts
rotate every 24 hours, but otherwise they should be persisted, so that
the user signatures generated throughout any given day for any given
`site_uuid` will be consistent.
- In production we use Firestore, and we have a simple MemoryStore
available for use in development. The MemoryStore is the default, and
therefore it's what is used in the self hoster configuration — but if
the container restarts, all the salts will be lost.
- This PR adds a better option for self hosters, which uses a simple
JSON file to persist the salts in a local file, which can be mounted
into the container.